Added gProduct Global Variable Model Fill upon inserting a new batch ; Corrected data type declaration on Docstamp Price from int to double (PriceListModel); Corrected  Naming convention from Bank to BankCode (PriceListModel); Added Product Code Field on History Schema (All Banks); Added Product code declaration on salesinvoice model field; Modified fetching product price to gProduct;

+++ CPMS-Accounting/Models/DocStampModel.cs
@@ -16,7 +16,7 @@ namespace CPMS_Accounting.Models
         public string DocDesc { get; set; }
         public string ChkType { get; set; }
         public int POorder { get; set; }
-        public int DocStampPrice { get; set; }
+        public double DocStampPrice { get; set; }
         public double TotalAmount { get; set; }
         public double unitprice { get; set; }
         public string BankCode { get; set; }
